fix File appender test on Win32
The test was reusing a file name in two test cases.  Because the log
file is still opened by the logger, it can't be unlinked on Win32.
The old file thus still has the content from the previous test case.
Change the second file name used to avoid this issue.